requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
youtube-search
python-multipart
pillow
aiofiles
